Adjust tests structure
