+++ tests/core/events/test_registry.py
