+++ tests/core/test_utils.py
